requirements.txt
pyyaml
snowflake-connector-python
permifrost>=0.15.0,<0.16.0